requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
Werkzeug==3.0.2
SQLAlchemy==2.0.29
gunicorn
psycopg2-binary
